zzform upload: zz_cleanup_dirs() stops if no handle was gotten by opendir()
git-svn-id: file:///Repos/zzform/trunk@1210 98be6e6d-ea4f-4117-9388-ed847deb6ed9

+++ zzform/upload.inc.php
@@ -2446,13 +2446,15 @@ function zz_cleanup_dirs($dir) {
 	$success = false;
 	if (is_dir($dir)) {
 		$dir_handle = opendir($dir);
-		$i = 0;
-		// check if directory is empty
-		while ($filename = readdir($dir_handle)) {
-			if ($filename != '.' AND $filename != '..') $i++;
+		if ($dir_handle) {
+			$i = 0;
+			// check if directory is empty
+			while ($filename = readdir($dir_handle)) {
+				if ($filename != '.' AND $filename != '..') $i++;
+			}
+			closedir($dir_handle);
+			if ($i == 0) $success = rmdir($dir);
 		}
-		closedir($dir_handle);
-		if ($i == 0) $success = rmdir($dir);
 	}
 	if ($success) {
 		// walk through dirs recursively
